test_requirements.txt
# Test Dependencies for Milvus Plugin Rewrite

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Mocking and fixtures
unittest-mock>=1.0.1
responses>=0.23.0
factory-boy>=3.2.0

# Test data generation
faker>=18.0.0

# Additional testing utilities
freezegun>=1.2.0  # For time-based testing
parameterized>=0.9.0  # For parameterized tests

# Development dependencies
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.0.0  # Type checking

# For integration tests
requests>=2.28.0
PyYAML>=6.0